move preview into it's own file, add format
